app/requirements.txt
Django==2.2.12
gunicorn==19.9.0
psycopg2-binary==2.8.3
djangorestframework==3.11.0
coverage>=5.1
django-sendgrid-v5==0.8.1
django-rest-swagger==2.2.0